Separate files to use modules system js best practices
