+++ src/index.html
@@ -22,9 +22,6 @@
                 <button class="py-2 px-3" id="refresh-button">Refresh</button>
             </div>
             <ul class="score-board" style="border: 1px solid black;">
-                <li>1</li>
-                <li>2</li>
-                <li>3</li>
             </ul>
         </div>
         <div class="col-6 add-score-section">
