+++ src/modules/returnGameId.js
@@ -1,0 +1,6 @@
+const returnGameId = (id) => {
+  const gameId = id;
+  return gameId;
+};
+
+export default returnGameId;
